+++ src/system/bashrc/ssh-agent.bashrc
@@ -1,7 +1,12 @@
 #!/bin/bash
-SSHAGENT=`which ssh-agent`
-SSHAGENTARGS="-s"
-if [ -z "$SSH_AUTH_SOCK" -a -x "$SSHAGENT" ]; then
-  eval `$SSHAGENT $SSHAGENTARGS`
-  trap "kill $SSH_AGENT_PID" 0
+
+if command_exists ssh-agent; then
+
+  SSHAGENT=`which ssh-agent`
+  SSHAGENTARGS="-s"
+  if [ -z "$SSH_AUTH_SOCK" -a -x "$SSHAGENT" ]; then
+    eval `$SSHAGENT $SSHAGENTARGS`
+    trap "kill $SSH_AGENT_PID" 0
+  fi
+
 fi
